New task item styling
